backend/requirements.txt
fastapi 
uvicorn 
redis 
httpx 
python-dotenv
pydantic